requirements.txt
appdirs==1.4.4
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
distlib==0.3.0
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pythonenv==0.0.34
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.6.0
virtualenv==20.0.21
Werkzeug==1.0.1
zipp==3.1.0
